Moved calculateSlippage from parsers to transformers
